\name{autism_data}
\docType{data}
\alias{autism_data}
\title{Autism Dataset}
\description{
Subset of a dataset of on the live of 3521 Dutch individuals diagnosed with Autism Disorder Spectrum (ASD). The full dataset can be found here https://github.com/jmbh/AutismData
}

\usage{data(autism_data)}

\value{
  \item{autism_data}{List containing the data, the type of variables, the number of categories of the variables and the column names.}
  }

\keyword{datasets}  

