\name{autism_data}
\docType{data}
\alias{autism_data}
\alias{autism_data_large}
\alias{mgm_data}
\alias{mvar_data}
\alias{restingstate_data}
\alias{fruitfly_data}
\alias{symptom_data}


\title{Autism Dataset}
\description{
FILL IN: description of and reference for all datasets


Subset of a dataset of on the live of 3521 Dutch individuals diagnosed with Autism Disorder Spectrum (ASD). The full dataset used in Deserno et al. (2016) can be found here https://github.com/jmbh/AutismData.
}

\usage{data(autism_data)}

\value{
  \item{autism_data}{List containing the data, the type of variables, the number of categories of the variables and the column names.}
  }

\references{
Deserno, M. K., Borsboom, D., Begeer, S., & Geurts, H. M. (2016). Multicausal systems ask for multicausal approaches: A network perspective on subjective well-being in individuals with autism spectrum disorder. Autism.
}

\keyword{datasets}

