\name{mgm-package}
\alias{mgm-package}
\alias{mgm-package}
\docType{package}
\title{
Estimating Time-Varying k-order Mixed Graphical Models
}
\description{
Estimation of time-varying Mixed Graphical models and mixed VAR models via elastic-net regularized neighborhood regression.
}
\details{
\tabular{ll}{
Package: \tab mgm\cr
Type: \tab Package\cr
Version: \tab 1.2-1 \cr
Date: \tab 2017-06-20\cr
License: \tab GPL-2 \cr
}
}
\author{
Jonas Haslbeck

Maintainer: <jonashaslbeck@gmail.com>
}


\references{

Haslbeck, J., & Waldorp, L. J. (2018). mgm: Estimating time-varying mixed graphical models in high-dimensional data. arXiv preprint http://arxiv.org/abs/1510.06871v2.

Haslbeck, J., & Waldorp, L. J. (2015). Structure estimation for mixed graphical models in high-dimensional data. arXiv preprint arXiv:1510.05677.

Loh, P. L., & Wainwright, M. J. (2013). Structure estimation for discrete graphical models: Generalized covariance matrices and their inverses. The Annals of Statistics, 41(6), 3022-3049.

Yang, E., Baker, Y., Ravikumar, P., Allen, G., & Liu, Z. (2014). Mixed graphical models via exponential families. In Proceedings of the Seventeenth International Conference on Artificial Intelligence and Statistics (pp. 1042-1050).
}

\keyword{package}
