\name{print.int}
\alias{print.int}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print method for int objects
}
\description{
Returns basic information about objects created with showInteraction()
}
\usage{
\method{print}{int}(x, ...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
The output object of showInteraction().
}
  \item{\dots}{
Additional arguments.
}
}
\value{
Writes basic information about the object in the console.
}

\author{
Jonas Haslbeck <jonashaslbeck@gmail.com>
}
