\name{print.rs}
\alias{print.rs}
\title{
Print method for rs objects}

\description{
Returns the model class, number of nodes, sample size of resamples, resampling with replacement \code{TRUE/FALSE} and iterations of resampling scheme.}

\usage{
\method{print}{rs}(x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{rs object (output of function \code{rs.mgm()})}
  \item{\dots}{Not used.}
}

\author{
Jonas Haslbeck <jonashaslbeck@gmail.com>}

