\name{mgm-package}
\alias{mgm-package}
\alias{mgm}
\docType{package}
\title{
Estimation of time-varying mixed graphical models and mixed vector autoregressive (VAR) models
}
\description{
Estimation of time-varying mixed graphical models and mixed vector autoregressive (VAR) models using L1-constrained neighborhood regression. 
}
\details{
\tabular{ll}{
Package: \tab mgm\cr
Type: \tab Package\cr
Version: \tab 1.1-6 \cr
Date: \tab 2016-04-27\cr
License: \tab GPL-2 \cr
}
}
\author{
Jonas Haslbeck

Maintainer: <jonashaslbeck@gmail.com>
}



\references{
Haslbeck, J., & Waldorp, L. J. (2015). Structure estimation for mixed graphical models in high-dimensional data. arXiv preprint arXiv:1510.05677.

Loh, P. L., & Wainwright, M. J. (2013). Structure estimation for discrete graphical models: Generalized covariance matrices and their inverses. The Annals of Statistics, 41(6), 3022-3049.

Yang, E., Baker, Y., Ravikumar, P., Allen, G., & Liu, Z. (2014). Mixed graphical models via exponential families. In Proceedings of the Seventeenth International Conference on Artificial Intelligence and Statistics (pp. 1042-1050).
}
\keyword{ package }
