\name{mgm-package}
\alias{mgm-package}
\alias{mgm-package}
\docType{package}
\title{
Estimating Time-Varying k-order Mixed Graphical Models
}
\description{
Estimation of time-varying Mixed Graphical models and mixed VAR models via elastic-net regularized neighborhood regression.
}
\details{
\tabular{ll}{
Package: \tab mgm\cr
Type: \tab Package\cr
Version: \tab 1.2-10 \cr
Date: \tab 2020-06-27\cr
License: \tab GPL-2 \cr
}
}
\author{
Jonas Haslbeck

Maintainer: <jonashaslbeck@gmail.com>
}


\references{

Haslbeck, J. M. B., & Waldorp, L. J. (2020). mgm: Estimating time-varying Mixed Graphical Models in high-dimensional Data. Journal of Statistical Software, 93(8), pp. 1-46. DOI: 10.18637/jss.v093.i08

Loh, P. L., & Wainwright, M. J. (2013). Structure estimation for discrete graphical models: Generalized covariance matrices and their inverses. The Annals of Statistics, 41(6), 3022-3049.

Yang, E., Baker, Y., Ravikumar, P., Allen, G., & Liu, Z. (2014). Mixed graphical models via exponential families. In Proceedings of the Seventeenth International Conference on Artificial Intelligence and Statistics (pp. 1042-1050).
}

\keyword{package}
