\name{mgm-package}
\alias{mgm-package}
\alias{mgm}
\docType{package}
\title{
Estimation of Mixed Graphical Models
}
\description{
Estimation of Mixed Graphical Models using L1-constrained neighborhood regression. 
}
\details{
\tabular{ll}{
Package: \tab mgm\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2015-07-08\cr
License: \tab ? \cr
}
}
\author{


Jonas Haslbeck

Maintainer: <jonashaslbeck@gmail.com>
}
\references{
Loh, P. L., & Wainwright, M. J. (2013). Structure estimation for discrete graphical models: Generalized covariance matrices and their inverses. The Annals of Statistics, 41(6), 3022-3049.

Yang, E., Baker, Y., Ravikumar, P., Allen, G., & Liu, Z. (2014). Mixed graphical models via exponential families. In Proceedings of the Seventeenth International Conference on Artificial Intelligence and Statistics (pp. 1042-1050).

}
\keyword{ package }
\examples{}
