\name{print.mgm}
\alias{print.mgm}
\title{
print a mgm object}
\description{
Print a selection of the output-list of the function \code{mgmfit}}
\usage{
\method{print}{mgm}(x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{fitted mgm object (output of function \code{mgmfit}.)}
  \item{\dots}{additional print arguments}
}
\value{
The adjacency matrix and the weighted adjacency matrix is printed}

\author{
Jonas Haslbeck <jonashaslbeck@gmail.com>}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
