\name{print.mgm}
\alias{print.mgm}
\title{
print a mgm object}
\description{
Print a selection of the output-list of the function \code{mgmfit}}
\usage{
\method{print}{mgm}(x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{fitted mgm object (output of function \code{mgmfit}.)}
  \item{\dots}{additional print arguments}
}
\value{
Prints the weighted adjacency matrix.}

\author{
Jonas Haslbeck <jonashaslbeck@gmail.com>}

